Ignore the tests which report leaks sporadically. This should cause few enough spurious warnings to send to python-checkins now.
If a test leaks that is not in the ignore list (LEAKY_TESTS), all tests
which leak will be reported.  This includes those in the ignore list.
Hopefully that will prompt someone to fix the flakiness.

--HG--
branch : legacy-trunk
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/legacy-trunk%4042500