SequenceMatcher.get_matching_blocks():  This now guarantees that adjacent triples in the result list describe non-adjacent matching blocks.  That's _nice_ to have, and Guido said he wanted it.
Not a bugfix candidate:  Guido or not ;-), this changes visible
endcase semantics (note that some tests had to change), and
nothing about this was documented before.  Since it was working
as designed, and behavior was consistent with the docs, it wasn't
"a bug".

--HG--
branch : legacy-trunk
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/legacy-trunk%4046939