Move fetching of encoding test files from the end of the Windows builbot's "build" step to the start of its "test" step.
This is poke-and-hope.  The hope is that compilation failures
on Windows will become visible to the buildbot (bsddb has
apparently been failing to compile in 2.4 on Windows "for
some time" now, but the buildbots haven't noticed that).

--HG--
branch : 2.4
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/release24-maint%4052259