Add an optional address to copy the failure mails to.
Detect a conflict in the only file that should have outstanding changes
when this script is run.  This doesn't matter on the trunk, but does
when run on a branch.  Trunk always has the date set to today in
boilerplate.tex.  Each time a release is cut with a different date,
a conflict occurs.  (We could copy a known good version, but then
we would lose changes to this file.)

--HG--
branch : legacy-trunk
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/legacy-trunk%4054884