A number of APIs in macostools cannot work in 64-bit mode because they use Carbon APIs that aren't available there.
This patch disables tests for the affected entrypoints in macostools and
mentions this in the documentation.

In theory it is possible to replace the implementation by code that does
work in 64-bit mode, but that would require further updates to the Carbon
wrappers because the modern APIs aren't wrapped properly.

--HG--
branch : legacy-trunk
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/legacy-trunk%4078061